+++ automaton/grid/_1d.hpp
@@ -8,18 +8,18 @@
 namespace automaton {
 
 struct cell_1d : public base_cell {
-    cell_1d(size_t row, size_t col);
+    cell_1d(uint32_t row, uint32_t col);
     bool operator<(const base_cell& other) const override;
 };
 
 class grid_1d : public base_grid {
 public:
-    grid_1d(size_t rows, size_t cols);
+    grid_1d(uint32_t rows, uint32_t cols);
     ~grid_1d() override;
 
-    void add(size_t row, size_t col) override;
-    bool remove(size_t row, size_t col) override;
-    bool has(size_t row, size_t col) override;
+    void add(uint32_t row, uint32_t col) override;
+    bool remove(uint32_t row, uint32_t col) override;
+    bool has(uint32_t row, uint32_t col) override;
     void clear() override;
 
     std::set<base_cell> get_drawable_cells() const override;
