+++ automaton/logic/wolfram.cpp
@@ -6,7 +6,7 @@ namespace logic {
 wolfram::wolfram(base_grid_ptr grid, int8_t code) : _grid(std::static_pointer_cast<grid_1d>(grid)), _code(code) {}
 
 void wolfram::step() {
-    size_t rows = _grid->get_rows(), cols = _grid->get_cols();
+    uint32_t rows = _grid->get_rows(), cols = _grid->get_cols();
 
     if (_current_step + 1 == rows) return;
 
@@ -14,11 +14,11 @@ void wolfram::step() {
     if (_current_step == 0 && _grid->get_data().size() > 0 && _grid->get_data().begin()->row != rows - 1)
         _current_step = _grid->get_data().begin()->row;
 
-    size_t row = _current_step;
+    uint32_t row = _current_step;
 
-    for (size_t col = 0; col < cols; col++) {
-        size_t prev_col = (col == 0 ? cols - 1 : col - 1);
-        size_t next_col = (col == cols - 1 ? 0 : col + 1);
+    for (uint32_t col = 0; col < cols; col++) {
+        uint32_t prev_col = (col == 0 ? cols - 1 : col - 1);
+        uint32_t next_col = (col == cols - 1 ? 0 : col + 1);
 
         int8_t combination = 0;
